Rename module `html_parser` to `feed_parser`
