+++ rssreader/tests.py
@@ -1,7 +1,7 @@
 import unittest
 
 from .storage.tests import *
-from .html_parser_tests import *
+from .feed_parser_tests import *
 from .output_controller_tests import *
 from .rss_reader_tests import *
